--- src/main/java/edu/rosehulman/covidtracer/controller/SystemUserController.java
@@ -1,7 +1,0 @@
-package edu.rosehulman.covidtracer.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class SystemUserController {
-}
