--- src/main/java/edu/rosehulman/covidtracer/model/SystemUser.java
@@ -1,59 +1,0 @@
-package edu.rosehulman.covidtracer.model;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.util.UUID;
-
-@Entity
-public class SystemUser implements Serializable {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	@Column(name = "ID", nullable = false)
-	private Long ID;
-
-	@OneToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-	@JoinColumn(name = "person_id", nullable = false, unique = true)
-	private Person person;
-
-	@Column(name = "Password", nullable = false)
-	private String password;
-
-	@Column(name = "Salt", nullable = false)
-	private String salt;
-
-	@ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-	@JoinColumn(name = "role_id", nullable = false)
-	private UserRole role;
-
-	public SystemUser() {
-	}
-
-	public SystemUser(Person person, String password, String salt, UserRole role) {
-		this.person = person;
-		this.password = password;
-		this.salt = salt;
-		this.role = role;
-	}
-
-	public Person getPerson() {
-		return person;
-	}
-
-	public Long ID() {
-		return ID;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public UserRole getRole() {
-		return role;
-	}
-
-	public String getSalt() {
-		return salt;
-	}
-
-}
